Renames variant folder names to match Braze; all fronts default to generic Collections
