+++ src/fronts/film-today/variantZ/Collections.tsx
