+++ src/fronts/film-today/variantZ/components/DefaultCollection.tsx
