+++ src/fronts/media-briefing/variantZ/components/CommentCollection.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Collection as ICollection } from "../../../../api";
 import { palette } from "@guardian/src-foundations";
-import { TableRowCell } from "../../../../layout/Table";
 import { Padding } from "../../../../layout/Padding";
 import { Multiline } from "../../../../components/Multiline";
 import { Heading } from "../../../../components/Heading";
@@ -15,12 +14,18 @@ export const CommentCollection: React.FC<{
     const white = palette.neutral[100];
 
     return (
-        <TableRowCell tdStyle={{ backgroundColor: lightGrey }}>
-            <Multiline topPadding />
-            <Heading heading={collection.displayName} />
+        <>
+            <Padding px={12} backgroundColor={lightGrey} />
+            <Multiline />
+            <Heading
+                heading={collection.displayName}
+                backgroundColor={lightGrey}
+            />
             {collection.backfill.map((story, index) => (
                 <>
-                    {index > 0 && <Padding px={4} />}
+                    {index > 0 && (
+                        <Padding px={12} backgroundColor={lightGrey} />
+                    )}
                     <HeadlineCard
                         content={story}
                         backgroundColor={white}
@@ -30,6 +35,6 @@ export const CommentCollection: React.FC<{
                     />
                 </>
             ))}
-        </TableRowCell>
+        </>
     );
 };
