+++ src/fronts/opinion/Opinion.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { Collection as ICollection } from "../../api";
+import { Collections } from "../../components/Collections";
 import { Collections as CollectionsVariantB } from "./variantB/Collections";
 import { Collections as CollectionsVariantC } from "./variantC/Collections";
 
@@ -17,13 +18,21 @@ export const Opinion: React.FC<{
                 salt={salt}
             ></CollectionsVariantC>
         );
+    } else if (variant === "b") {
+        return (
+            <CollectionsVariantB
+                frontId={frontId}
+                collections={collections}
+                salt={salt}
+            ></CollectionsVariantB>
+        );
     }
 
     return (
-        <CollectionsVariantB
+        <Collections
             frontId={frontId}
             collections={collections}
             salt={salt}
-        ></CollectionsVariantB>
+        ></Collections>
     );
 };
